rename: uiからpresentationに名前を変更
